+++ db/migrate/20130321141401_rename_librarian_to_editor.rb
@@ -1,0 +1,5 @@
+class RenameLibrarianToEditor < ActiveRecord::Migration
+  def change
+    rename_column :project_users, :librarian, :editor
+  end
+end