+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.15.0
 
+### Enhancements
+- Project users now simplified to editors and viewers, formerly librarians and members
+  - Viewers (formerly members) can no longer create or edit sheets
+
 ## 0.14.4 (March 20, 2013)
 
 ### Enhancements